# internal/transport/http/swaggerdocs/swagger.yaml
basePath: /api/v1
definitions:
...
  shared.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  shared.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/shared.ErrorBody'
    type: object
...
  shared.EventHonorBonusHonorGroupResponse:
    properties:
      backgroundAssetbundleName:
        type: string
      frameName:
        type: string
      honorType:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  shared.EventHonorBonusHonorResponse:
    properties:
      assetbundleName:
        type: string
      group:
        $ref: '#/definitions/shared.EventHonorBonusHonorGroupResponse'
      groupId:
        type: integer
      honorMissionType:
        type: string
      honorRarity:
        type: string
      honorTypeId:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  shared.EventHonorBonusLeaderGameCharacterUnitResponse:
    properties:
      gameCharacterId:
        type: integer
      id:
        type: integer
      unit:
        type: string
    type: object
  shared.EventHonorBonusListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/shared.EventHonorBonusObjectResponse'
        type: array
    type: object
  shared.EventHonorBonusObjectResponse:
    properties:
      bonusRate:
        type: integer
      eventId:
        type: integer
      honor:
        $ref: '#/definitions/shared.EventHonorBonusHonorResponse'
      honorId:
        type: integer
      id:
        type: integer
      leaderGameCharacterId:
        type: integer
      leaderGameCharacterUnit:
        $ref: '#/definitions/shared.EventHonorBonusLeaderGameCharacterUnitResponse'
    type: object
...
info:
  contact: {}
  description: API for master data sync and card querying.
  title: sekai-master-api
  version: "1.0"
paths:
...
  /events/{region}/{id}/honor-bonuses:
    get:
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.EventHonorBonusListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Get typed event honor bonuses by event id
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"